adjustments for detection
